Adds mariobehling to mentors section.
Adds a mentor to the mentors section and formatted it down.

The mentors section was statically generated which can and will lead to
pretty bad development scenarios as a large no of students will add more
than hundreds of mentors to the site. So this commit makes jekyll to
build the site from the data files which will contain mentor data.
Documentation on how to add a mentor has been added to readme.md.
In addition to the above mariobehling has been added to the site.

Signed-off-by: Abishek V Ashok <abishekvashok@gmail.com>