Adds support for Travis CI
Adds .travis.yml and Gemfile.
.travis.yml will be used to specify to Travis CI what should be run and
also what the build environment should be. Since, Jekyll works as a ruby
gem the language will be ruby and we will test that Jekyll can build the
website. The Gemfile contains dependency information for bundler to
fetch during bundler install so that jekyll could be executed.

Signed-off-by: Abishek V Ashok <abishekvashok@gmail.com>